TINY-9829: Rename EventUtils to PasteEventUtils
