+++ modules/tinymce/src/core/test/ts/module/test/PasteEventUtils.ts
